==> 01-intro.qmd <==
# Brief Introduction to Analyses

Typically, youth follow an expected growth curve, with gains in weight and height progressing in a **predictable, but individualized** manner reflecting previous development. In some cases, deviations from expected growth signal mental health concern. Historically, conceptualizations of restrictive-spectrum eating disorders (specifically anorexia nervosa) has included weight-related diagnostic markers informed by population-based norms; however, a large and growing body of research supports *individualized* markers of weight status (i.e. weight suppression -- the difference between current and highest past weight status -- which can be operationalized as deviation from expected growth amongst youth) as a key predictor of both risk and recovery for restrictive EDs. At present, use of individualized measurements to screen for and treat restrictive EDs is not widespread.

The current analyses aim to define a model which predicts expected growth (along with an expected window of BMI) for an individual across adolescence, based on data obtained from their height and weight throughout childhood. We will use this predicted model to examine whether a drop from expected BMI z-score during adolescence is likely to predict or be accompanied by eating disordered behaviors in the ALSPAC Cohort.
