==> 02-forecasting.qmd <==
# Forecasting BMI

```{r, setup, include = 'false', results = 'hide'}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE, out.width = '100%')
library(fable)
library(ggplot2)
library(dplyr)
library(stringr)
library(gridExtra)
library(haven)
library(zscorer)
library(psych)
library(AGD)
library(readxl)
load('data/ALSPAC_Cleaned.RData') 
load('data/fcast_data.RData')
source('R/bmiz_to_bmi.R')
source('R/solve_for_wt.R')
source('R/logit_transform.R')
source('R/add_bmi_vars.R')
```

There are `r count(bmi_df_u13 |> distinct(id))[[1,1]]` individuals with four or more bmi measurements in the dataset between ages 2-13 years.

```{r, include = 'false', results='hide', warning=FALSE}
sex_vars <- bmi_df |> 
  select(id, sex_z) |> 
  distinct()
fcast <- left_join(fcast_raw, sex_vars)  # adds sex as a variable to the forecasted datasets
fcast <- add_bmi_vars(fcast) #adds bmi variables to the forecasted dataset, including 95 and 99% forecasted ranges from each model that was run

#This creates a dataset with actual bmi values from age 13 plus
bmi_df_13_plus <- bmi_df |> 
  filter(agemos_asess_1 >= 13*12)

#This joins the actual recorded values from age 13+ for bmi with the forcasted values and ranges
bmi_df_13_plus <- merge(fcast, bmi_df_13_plus, by = c('id', 'agemos_asess_1', 'sex_z'))

# Makes a dataset with each of the models independently for future use
data_by_model <- split(bmi_df_13_plus, bmi_df_13_plus$.model)

```

When combining data with assessments of body weight after age 12.99, there were a total of `r nrow(data_by_model[[1]])` observations of body weight from ages 13-20 in the dataset, across `r count(data_by_model[[1]] |> distinct(id))[[1,1]]` participants.

```{r, include = FALSE, results = 'hide', warning=FALSE}
bmi_13_14 <- bmi_df_13_plus |> 
  filter(!is.na(bmiz)) |>
  filter(agemos_asess_1 < 14.0*12) |> #selects only assessment points between 13.0-14.0 years of age
  mutate(lower_99_cutoff =  case_when(bmiz < bmiz_lower_99 ~ 1, 
                                      bmiz >=bmiz_lower_99 ~ 0),
         lower_95_cutoff =  case_when(bmiz < bmiz_lower_95 ~ 1, 
                                      bmiz >=bmiz_lower_95 ~ 0),
          grade_1_cutoff =  case_when(bmiz < -1 ~ 1, 
                                      bmiz > -1 ~ 0),
          upper_99_cutoff =  case_when(bmiz > bmiz_upper_99 ~ 1, 
                                      bmiz < bmiz_upper_99 ~ 0),
          upper_95_cutoff =  case_when(bmiz > bmiz_upper_95 ~ 1, 
                                      bmiz <=bmiz_upper_95 ~ 0),
          overweight_cutoff =  case_when(bmiz > 1 ~ 1, 
                                      bmiz < 1 ~ 0)) |> 
  select(c(id, sex_z, agemos_asess_1, assess_type, bmiz, bmiz_lower_99, bmiz_upper_99, lower_99_cutoff, lower_95_cutoff, grade_1_cutoff, upper_99_cutoff, upper_95_cutoff, overweight_cutoff, .model))  |> 
  mutate(low_wt_category = factor(case_when(lower_99_cutoff == 0 & grade_1_cutoff == 0 ~ '0.No Low Wt',
                                     lower_99_cutoff == 0 & grade_1_cutoff == 1 ~ '2.UW - No BMIZ change',
                                     lower_99_cutoff == 1 & grade_1_cutoff == 0 ~  '1.BMIZ Reduce - No UW',
                                     lower_99_cutoff == 1 & grade_1_cutoff == 1 ~ '3.UW plus BMIZ Reduce'))) |> 
   mutate(high_wt_category = factor(case_when(upper_99_cutoff == 0 & overweight_cutoff == 0 ~ '0.No High Wt',
                                     upper_99_cutoff == 0 & overweight_cutoff == 1 ~ '2.OW - No BMIZ Elevation',
                                     upper_99_cutoff == 1 & overweight_cutoff == 0 ~ '1.BMIZ Elevate - No OW',
                                     upper_99_cutoff == 1 & overweight_cutoff == 1 ~ '3.BMIZ Elevate + OW'))) 
 
save(bmi_13_14, sex_vars, file = 'data/forecast_data.RData')    
```

```{r, include = FALSE, results = 'hide', warning=FALSE}
bmi_13_16 <- bmi_df_13_plus |> 
  filter(!is.na(bmiz)) |>
  filter(agemos_asess_1 < 16*12) |> #selects only assessment points between 13.0-14.0 years of age
  mutate(lower_99_cutoff =  case_when(bmiz < bmiz_lower_99 ~ 1, 
                                      bmiz >=bmiz_lower_99 ~ 0),
         lower_95_cutoff =  case_when(bmiz < bmiz_lower_95 ~ 1, 
                                      bmiz >=bmiz_lower_95 ~ 0),
          grade_1_cutoff =  case_when(bmiz < -1 ~ 1, 
                                      bmiz > -1 ~ 0),
          upper_99_cutoff =  case_when(bmiz > bmiz_upper_99 ~ 1, 
                                      bmiz < bmiz_upper_99 ~ 0),
          upper_95_cutoff =  case_when(bmiz > bmiz_upper_95 ~ 1, 
                                      bmiz <=bmiz_upper_95 ~ 0),
          overweight_cutoff =  case_when(bmiz > 1 ~ 1, 
                                      bmiz < 1 ~ 0)) |> 
  select(c(id, sex_z, agemos_asess_1, assess_type, bmiz, bmiz_lower_99, bmiz_upper_99, lower_99_cutoff, lower_95_cutoff, grade_1_cutoff, upper_99_cutoff, upper_95_cutoff, overweight_cutoff, .model))  |> 
  mutate(low_wt_category = factor(case_when(lower_99_cutoff == 0 & grade_1_cutoff == 0 ~ '0.No Low Wt',
                                     lower_99_cutoff == 0 & grade_1_cutoff == 1 ~ '2.UW - No BMIZ change',
                                     lower_99_cutoff == 1 & grade_1_cutoff == 0 ~  '1.BMIZ Reduce - No UW',
                                     lower_99_cutoff == 1 & grade_1_cutoff == 1 ~ '3.UW plus BMIZ Reduce'))) |> 
   mutate(high_wt_category = factor(case_when(upper_99_cutoff == 0 & overweight_cutoff == 0 ~ '0.No High Wt',
                                     upper_99_cutoff == 0 & overweight_cutoff == 1 ~ '2.OW - No BMIZ Elevation',
                                     upper_99_cutoff == 1 & overweight_cutoff == 0 ~ '1.BMIZ Elevate - No OW',
                                     upper_99_cutoff == 1 & overweight_cutoff == 1 ~ '3.BMIZ Elevate + OW'))) 
 
save(bmi_13_16, sex_vars, file = 'data/bmi_13_16.RData')    
```
